+++ frontend/src/api/mypage/orderApi.ts
@@ -1,17 +1,8 @@
 import axios from "../axios";
-
-export interface MypageOrdersListDto {
-  orderId: number;
-  orderDt: string;
-  orderStatus: number;
-  orderTravelId: number;
-  orderTravelName: string;
-  orderTravelAmount: number;
-  travelStartDt: string;
-  travelEndDt: string;
-  travelImg: string;
-  totalPrice: number;
-}
+import {
+  MypageOrdersListDto,
+  MypageOrdersReviewDto,
+} from "../../types/mypage/order";
 
 export const getOrdersList = async (
   memberId: string,
@@ -24,11 +15,6 @@ export const getOrdersList = async (
   return response.data;
 };
 
-export interface MypageOrdersUpdDto {
-  orderId: number;
-  orderStatus: number;
-}
-
 export const cancelOrder = async (orderId: number): Promise<string> => {
   const response = await axios.put("/mypage/orders/updOrdersCancel", {
     orderId,
@@ -36,13 +22,6 @@ export const cancelOrder = async (orderId: number): Promise<string> => {
   });
   return response.data;
 };
-
-export interface MypageOrdersReviewDto {
-  travelId: number;
-  memberId: string;
-  reviewRate: number;
-  reviewComment: string;
-}
 
 export const addOrdersReview = async (
   review: MypageOrdersReviewDto,
