+++ frontend/src/types/mypage/member.ts
@@ -1,0 +1,5 @@
+export interface MypageMemberProfileDto {
+  memberId: string;
+  memberName: string;
+  memberImg: string;
+}
